+++ amqpstorm/tests/functional/__init__.py
@@ -1,0 +1,4 @@
+HOST = '127.0.0.1'
+USERNAME = 'guest'
+PASSWORD = 'guest'
+URI = 'amqp://guest:guest@127.0.0.1:5672/%2F'