+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
 Changelog
 -------------
 
+Version 2.0.0
+-------------
+- Messages are now delivered as Message objects by default.
+    - to_tuple and to_dict are now set to False by default.
+
+        This is a breaking change that affects the following function:
+
+            - channel.process_data_events
+            - channel.start_consuming
+            - channel.basic.get
+
 Version 1.5.0
 -------------
 - Added support for Channel.Tx (Server local transactions). [#27]